Validations, exception handling and bcrypt password done

+++ src/main/java/com/maveric/userservice/service/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.maveric.userservice.service;
 
 import com.maveric.userservice.dto.UserResponse;
+import com.maveric.userservice.exceptionhandler.EmailAlreadyExist;
 import com.maveric.userservice.exceptionhandler.EmailNotFound;
 import com.maveric.userservice.exceptionhandler.UserNotExist;
 import com.maveric.userservice.mapper.UserMapper;
@@ -11,7 +12,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
+import org.springframework.security.crypto.password.PasswordEncoder;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,6 +24,9 @@ public class UserServiceImpl implements UserService {
 
     @Autowired
     private UserMapper mapper;
+
+    @Autowired
+    PasswordEncoder passwordEncoder;
     public List<UserResponse> getUsers(Integer page, Integer pageSize) {
         Pageable paging = (Pageable) PageRequest.of(page, pageSize);
         Page<User> pageResult = repository.findAll(paging);
@@ -38,11 +42,19 @@ public class UserServiceImpl implements UserService {
         }
     }
     public UserResponse createUser(UserResponse userResponse) {
-        //Adding CreatedTime
+//        User user = mapper.map(userResponse);
+//        User userResult = repository.save(user);
+//        return  mapper.map(userResult);
 
+
+
+        String pass = passwordEncoder.encode(userResponse.getPassword());
+        userResponse.setPassword(pass);
         User user = mapper.map(userResponse);
         User userResult = repository.save(user);
         return  mapper.map(userResult);
+
+
     }
 
     @Override
@@ -53,8 +65,14 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public String deleteUser(String userId) {
-        repository.deleteById(Long.valueOf(userId));
-        return "The user is deleted successfully.";
+
+        if(repository.findById(Long.valueOf(userId)).isPresent()){
+            repository.deleteById(Long.valueOf(userId));
+        }
+        else {
+            throw new UserNotExist("No Users Found");
+        }
+        return "User Successfully Deleted";
     }
 
     public UserResponse getUserDetailsByEmail(String email) {
@@ -73,7 +91,16 @@ public class UserServiceImpl implements UserService {
         updateUser.setLastName(user.getLastName());
         updateUser.setMiddleName(user.getMiddleName());
         updateUser.setPhoneNumber(user.getPhoneNumber());
-        updateUser.setEmail(user.getEmail());
+//        updateUser.setEmail(user.getEmail());
+
+
+        if (repository.findByEmail(user.getEmail())!=null) {
+            throw new EmailAlreadyExist("Email Already Exist");
+        }
+        else{
+            updateUser.setEmail(user.getEmail());
+        }
+
         updateUser.setAddress(user.getAddress());
         updateUser.setDateOfBirth(user.getDateOfBirth());
         updateUser.setGender(user.getGender());
